+++ sdk/resourcemanager/compute/armcompute/responses.go
@@ -1023,6 +1023,12 @@ type SnapshotsClientUpdateResponse struct {
 	Snapshot
 }
 
+// SoftDeleteClientListSoftDeletedResourcesByArtifactNameResponse contains the response from method SoftDeleteClient.NewListSoftDeletedResourcesByArtifactNamePager.
+type SoftDeleteClientListSoftDeletedResourcesByArtifactNameResponse struct {
+	// The List Soft-deleted Resources operation response.
+	GallerySoftDeletedResourceList
+}
+
 // UsageClientListResponse contains the response from method UsageClient.NewListPager.
 type UsageClientListResponse struct {
 	// The List Usages operation response.